Clean up client engine specs
